pyproject.toml
[project]
name = "pa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.24.0",
    "requests>=2.32.5",
    "tavily-python>=0.7.22",
    "python-dotenv>=1.0.1",
]
